requirements.txt
#pip install tortoise-orm[aiosqlite]
tortoise-orm==0.17.3
aiosqlite==0.16.1
